recommendations/requirements.txt
grpcio-tools==1.56.0
pytest==7.4.0
grpc-interceptor==0.15.2
